misc: qcom: qdsp6v2: Do not truncate the buf address
Truncating the buffer address would cause unmap failure
and hence next session playback would fail eventually.
Do not truncate the buffer address which is needed for
the memory unmap.

CRs-Fixed: 814704, 814699

+++ drivers/misc/qcom/qdsp6v2/audio_utils_aio.c
@@ -1031,7 +1031,7 @@ static int audio_aio_ion_remove(struct q6audio_aio *audio,
 			pr_debug("%s[%pK]:remove region fd %d vaddr %pK\n",
 				__func__, audio, info->fd, info->vaddr);
 			rc = q6asm_memory_unmap(audio->ac,
-						(uint32_t) region->paddr, IN);
+						region->paddr, IN);
 			if (rc < 0)
 				pr_err("%s[%pK]: memory unmap failed\n",
 					__func__, audio);
